alteração no calculo

+++ Models/Piso.cs
@@ -6,10 +6,10 @@
         public string Local { get; set; }
         public int Area1 { get; set; }
         public int Area2 { get; set; }
-        public int Piso1 { get; set; }
-        public int Piso2 { get; set; }
-        public int CalculoArea { get; set; }
-        public int CalculoPiso { get; set; }
+        public float Piso1 { get; set; }
+        public float Piso2 { get; set; }
+        public float CalculoArea { get; set; }
+        public float CalculoPiso { get; set; }
 
 
     }
